Impute values from prototype information

Impute values by replacing missing entries with the corresponding assigned prototype entries

impute(object, ...)

Arguments

object

a somRes object.

...

unused.

Value

Imputed matrix as in Cottrell and Letrémy, (2005)

References

Cottrell M., Letrémy P. (2005) Missing values: processing with the Kohonen algorithm. Proceedings of Applied Stochastic Models and Data Analysis (ASMDA 2005), 489-496.

See also

Author

Nathalie Vialaneix nathalie.vialaneix@inrae.fr

Examples

# Run trainSOM algorithm on the iris data with 500 iterations
set.seed(1505)
missings <- cbind(sample(1:150, 50, replace = TRUE),
                  sample(1:4, 50, replace = TRUE))
x.data <- as.matrix(iris[, 1:4])
x.data[missings] <- NA
iris.som <- trainSOM(x.data = x.data)
iris.som
#>       Self-Organizing Map object...
#>          online learning, type: numeric 
#>          5 x 5 grid with square topology
#>          neighbourhood type: gaussian 
#>          distance type: euclidean 
impute(iris.som)
#>        Sepal.Length Sepal.Width Petal.Length Petal.Width
#>   [1,]     5.100000    3.500000     1.400000   0.2000000
#>   [2,]     4.900000    3.000000     1.400000   0.2000000
#>   [3,]     4.700000    3.200000     1.300000   0.2000000
#>   [4,]     4.600000    3.100000     1.500000   0.2000000
#>   [5,]     5.000000    3.600000     1.400000   0.2000000
#>   [6,]     5.400000    3.900000     1.700000   0.4000000
#>   [7,]     4.600000    3.291352     1.400000   0.3000000
#>   [8,]     5.000000    3.400000     2.066971   0.5112279
#>   [9,]     4.400000    2.900000     1.400000   0.2000000
#>  [10,]     4.900000    3.100000     1.500000   0.1000000
#>  [11,]     5.400000    3.291352     1.500000   0.2000000
#>  [12,]     4.800000    3.400000     1.600000   0.2000000
#>  [13,]     4.800000    3.000000     1.400000   0.1000000
#>  [14,]     4.300000    3.000000     1.100000   0.1000000
#>  [15,]     5.800000    4.000000     1.200000   0.2000000
#>  [16,]     5.700000    4.400000     1.500000   0.4000000
#>  [17,]     5.400000    3.900000     1.300000   0.4000000
#>  [18,]     5.100000    3.500000     1.400000   0.3000000
#>  [19,]     5.700000    3.800000     1.700000   0.3000000
#>  [20,]     5.100000    3.800000     1.500000   0.3000000
#>  [21,]     5.400000    3.400000     1.700000   0.2000000
#>  [22,]     5.100000    3.700000     1.500000   0.5112279
#>  [23,]     4.600000    3.600000     1.000000   0.5112279
#>  [24,]     5.100000    3.300000     1.700000   0.5000000
#>  [25,]     5.168237    3.400000     1.900000   0.2000000
#>  [26,]     5.000000    3.000000     1.600000   0.2000000
#>  [27,]     5.000000    3.400000     1.600000   0.4000000
#>  [28,]     5.200000    3.500000     1.500000   0.2000000
#>  [29,]     5.200000    3.400000     1.400000   0.2000000
#>  [30,]     4.700000    3.200000     1.600000   0.2000000
#>  [31,]     5.168237    3.100000     2.066971   0.2000000
#>  [32,]     5.400000    3.400000     1.500000   0.4000000
#>  [33,]     5.200000    3.291352     1.500000   0.1000000
#>  [34,]     5.500000    4.200000     1.400000   0.2000000
#>  [35,]     4.900000    3.100000     1.500000   0.2000000
#>  [36,]     5.168237    3.200000     1.200000   0.2000000
#>  [37,]     5.500000    3.500000     1.300000   0.2000000
#>  [38,]     4.900000    3.291352     1.400000   0.1000000
#>  [39,]     4.400000    3.000000     1.300000   0.2000000
#>  [40,]     5.100000    3.400000     1.500000   0.2000000
#>  [41,]     5.000000    3.500000     1.300000   0.3000000
#>  [42,]     4.500000    2.300000     1.300000   0.3000000
#>  [43,]     4.400000    3.200000     1.300000   0.2000000
#>  [44,]     5.000000    3.500000     1.600000   0.6000000
#>  [45,]     5.100000    3.800000     1.900000   0.4000000
#>  [46,]     4.800000    3.000000     1.400000   0.3000000
#>  [47,]     5.100000    3.800000     1.600000   0.5112279
#>  [48,]     4.600000    3.200000     1.400000   0.2000000
#>  [49,]     5.300000    3.700000     1.500000   0.2000000
#>  [50,]     5.000000    3.300000     1.400000   0.2000000
#>  [51,]     7.000000    3.200000     4.700000   1.8439166
#>  [52,]     6.400000    3.200000     4.500000   1.5000000
#>  [53,]     6.131118    3.100000     4.362574   1.5000000
#>  [54,]     5.500000    2.300000     4.000000   1.3000000
#>  [55,]     6.500000    3.070049     4.600000   1.5000000
#>  [56,]     5.700000    2.800000     4.500000   1.3000000
#>  [57,]     6.300000    3.300000     4.700000   1.6000000
#>  [58,]     4.900000    2.400000     3.300000   1.3216226
#>  [59,]     6.600000    2.900000     4.600000   1.3000000
#>  [60,]     5.200000    2.700000     3.900000   1.4000000
#>  [61,]     5.000000    2.000000     3.500000   1.0000000
#>  [62,]     5.900000    3.000000     4.200000   1.5000000
#>  [63,]     6.000000    2.200000     4.000000   1.0000000
#>  [64,]     6.100000    2.949623     4.700000   1.4000000
#>  [65,]     5.600000    2.900000     3.600000   1.3000000
#>  [66,]     6.700000    3.100000     4.400000   1.4000000
#>  [67,]     6.103554    3.000000     4.500000   1.5627189
#>  [68,]     5.800000    2.700000     4.100000   1.0000000
#>  [69,]     5.742542    2.200000     4.500000   1.5000000
#>  [70,]     5.742542    2.500000     4.054982   1.1000000
#>  [71,]     5.900000    3.200000     4.800000   1.8000000
#>  [72,]     6.100000    2.800000     4.000000   1.3000000
#>  [73,]     6.300000    2.500000     4.900000   1.5000000
#>  [74,]     6.100000    2.800000     4.700000   1.2000000
#>  [75,]     6.400000    2.900000     4.300000   1.6443559
#>  [76,]     6.600000    3.000000     4.400000   1.4000000
#>  [77,]     6.800000    2.800000     4.800000   1.8439166
#>  [78,]     6.700000    3.000000     5.000000   1.7000000
#>  [79,]     6.000000    2.900000     4.500000   1.5000000
#>  [80,]     5.700000    2.600000     3.500000   1.0000000
#>  [81,]     5.500000    2.400000     3.800000   1.1000000
#>  [82,]     5.500000    2.883555     3.700000   1.0000000
#>  [83,]     5.800000    2.700000     3.900000   1.2000000
#>  [84,]     6.000000    2.700000     5.100000   1.6000000
#>  [85,]     5.400000    2.778708     4.500000   1.5000000
#>  [86,]     6.000000    3.400000     4.500000   1.6000000
#>  [87,]     6.700000    3.100000     4.700000   1.5000000
#>  [88,]     6.300000    2.300000     4.400000   1.4817026
#>  [89,]     5.600000    3.000000     3.417475   1.3000000
#>  [90,]     5.500000    2.500000     4.000000   1.3000000
#>  [91,]     5.500000    2.600000     4.400000   1.2000000
#>  [92,]     6.100000    3.000000     4.600000   1.4000000
#>  [93,]     5.800000    2.600000     4.000000   1.2000000
#>  [94,]     5.000000    2.300000     3.300000   1.0000000
#>  [95,]     5.600000    2.700000     4.200000   1.3000000
#>  [96,]     5.700000    3.000000     4.200000   1.2000000
#>  [97,]     5.700000    2.900000     4.200000   1.3000000
#>  [98,]     6.200000    2.900000     4.300000   1.3000000
#>  [99,]     5.100000    3.038599     3.000000   1.1000000
#> [100,]     5.700000    2.800000     4.100000   1.3000000
#> [101,]     6.300000    3.300000     6.000000   2.5000000
#> [102,]     5.800000    2.700000     5.100000   1.9000000
#> [103,]     7.100000    3.000000     5.900000   2.1000000
#> [104,]     6.300000    2.900000     5.600000   1.8000000
#> [105,]     6.500000    3.000000     5.800000   2.2000000
#> [106,]     7.600000    3.000000     6.600000   2.1000000
#> [107,]     4.900000    2.883555     4.500000   1.1638338
#> [108,]     7.300000    2.900000     6.300000   1.8000000
#> [109,]     6.700000    2.500000     5.800000   1.8000000
#> [110,]     7.200000    3.600000     6.100000   2.5000000
#> [111,]     6.500000    3.200000     5.100000   2.0000000
#> [112,]     6.125973    2.700000     5.300000   1.6443559
#> [113,]     6.455103    3.000000     5.500000   2.1000000
#> [114,]     5.700000    2.500000     5.000000   1.3216226
#> [115,]     5.800000    2.800000     5.100000   2.4000000
#> [116,]     6.400000    3.200000     5.300000   2.3000000
#> [117,]     6.500000    3.000000     5.500000   1.8000000
#> [118,]     6.455103    3.800000     6.700000   2.2000000
#> [119,]     7.700000    2.600000     5.150084   2.3000000
#> [120,]     6.000000    2.200000     5.000000   1.5000000
#> [121,]     6.900000    3.200000     5.700000   2.3000000
#> [122,]     5.600000    2.893430     4.900000   2.0000000
#> [123,]     7.700000    2.800000     6.700000   2.0000000
#> [124,]     6.300000    2.700000     4.900000   1.8000000
#> [125,]     6.700000    3.021305     5.700000   2.1000000
#> [126,]     7.200000    3.200000     6.000000   1.8000000
#> [127,]     6.200000    2.800000     4.800000   1.8000000
#> [128,]     6.100000    3.000000     4.900000   1.8000000
#> [129,]     6.455103    2.800000     5.600000   2.1000000
#> [130,]     7.200000    3.000000     5.800000   1.6000000
#> [131,]     7.400000    2.800000     6.100000   1.9000000
#> [132,]     7.900000    3.800000     6.400000   2.0000000
#> [133,]     6.400000    2.800000     5.600000   2.2000000
#> [134,]     6.300000    2.800000     4.683201   1.5000000
#> [135,]     6.100000    2.600000     5.600000   1.4000000
#> [136,]     7.700000    3.000000     6.100000   2.3000000
#> [137,]     6.300000    3.400000     5.600000   2.4000000
#> [138,]     6.400000    3.100000     5.500000   1.8000000
#> [139,]     6.000000    3.000000     4.771238   1.8000000
#> [140,]     6.900000    3.100000     5.400000   2.1000000
#> [141,]     6.455103    3.100000     5.600000   2.4000000
#> [142,]     6.900000    3.100000     5.100000   2.3000000
#> [143,]     5.800000    2.700000     5.100000   1.9000000
#> [144,]     6.800000    3.200000     5.150084   2.3000000
#> [145,]     6.700000    3.300000     5.700000   2.5000000
#> [146,]     6.700000    3.000000     5.150084   2.3000000
#> [147,]     6.300000    2.500000     5.000000   1.4817026
#> [148,]     6.500000    3.000000     5.200000   2.0000000
#> [149,]     6.200000    3.400000     5.005328   2.3000000
#> [150,]     5.900000    3.000000     4.400976   1.8000000