Dataset "Les Misérables"

This dataset contains the coappearance network (igraph object) of characters in the novel Les Misérables (written by the French writter Victor Hugo).

Format

lesmis is an igraph object. Its vertices are the characters of the novel and an edge indicates that the two characters appear together in the same chapter of the novel, at least once. Vertex attributes for this graph are id, a vertex number between 1 and 77, and label, the character's name. The edge attribute value gives the number of co-appearances between the two characters afferent to the edge (the igraph can thus be made a weighted graph using this attribute). Finally, a graph attribute layout is used to provide a layout (generated with the igraph function layout_with_fr) for visualizing the graph.

dissim.lesmis is a dissimilarity matrix computed with the function shortest_paths and containing the length of the shortest paths between pairs of nodes.

Details

Les Misérables is a French historical novel, written by Victor Hugo and published in 1862. The co-appearance network has been extracted by D.E. Knuth (1993).

References

Hugo V. (1862) Les Miserables.

Knuth D.E. (1993) The Stanford GraphBase: A Platform for Combinatorial Computing. Reading (MA): Addison-Wesley.

Examples

data(lesmis)
if (FALSE) {
summary(lesmis)
plot(lesmis,vertex.size=0)}