Compute distances between prototypes

Compute distances, either between all prototypes (mode="complete") or only between prototypes' neighbours (mode="neighbors").

# S3 method for somRes
protoDist(object, mode=c("complete","neighbors"), ...)

Arguments

object
a somRes object.

mode
Specifies which distances should be computed.

...
Not used.

Details

When mode="complete", distances between all prototypes are computed. When mode="neighbors", distances are computed only between the prototypes and their neighbors. If the data were preprocessed during the SOM training procedure, the distances are computed on the normalized values of the prototypes.

Value

When mode="complete", the function returns a square matrix which dimensions are equal to the product of the grid dimensions.

When mode="neighbors", the function returns a list which length is equal to the product of the grid dimensions; the length of each item is equal to the number of neighbors. Neurons are considered to have 8 neighbors at most (i.e., two neurons are neighbors if they have a distance of type 'maximum' which is equal to 1).

See also

Examples

set.seed(2343)
my.som <- trainSOM(x.data=iris[,1:4], dimension=c(5,5))
protoDist(my.som)

#>            1         2         3         4         5         6         7
#> 1  0.0000000 0.4153914 0.7501372 0.9910827 1.1339710 0.4762939 0.0697045
#> 2  0.4153914 0.0000000 0.3347533 0.5757144 0.7186563 0.8912676 0.4003482
#> 3  0.7501372 0.3347533 0.0000000 0.2409746 0.3839940 1.2258708 0.7325264
#> 4  0.9910827 0.5757144 0.2409746 0.0000000 0.1431262 1.4667063 0.9725654
#> 5  1.1339710 0.7186563 0.3839940 0.1431262 0.0000000 1.6094271 1.1147812
#> 6  0.4762939 0.8912676 1.2258708 1.4667063 1.6094271 0.0000000 0.4975065
#> 7  0.0697045 0.4003482 0.7325264 0.9725654 1.1147812 0.4975065 0.0000000
#> 8  0.4685045 0.1197270 0.3163142 0.5477014 0.6871815 0.9339559 0.4365580
#> 9  0.7755702 0.3778043 0.1551069 0.2768914 0.4013983 1.2418792 0.7446475
#> 10 0.9500087 0.5488347 0.2600959 0.1933986 0.2689920 1.4158657 0.9188542
#> 11 1.0855087 1.5006379 1.8352388 2.0760548 2.2187379 0.6094186 1.1055983
#> 12 0.5580196 0.9665308 1.2988762 1.5385417 1.6803898 0.1237082 0.5664829
#> 13 0.2481909 0.4757039 0.7810246 1.0118296 1.1494993 0.5323894 0.1818911
#> 14 0.5429496 0.3489695 0.4830211 0.6713901 0.7942539 0.9474151 0.4859885
#> 15 0.7630428 0.4729018 0.4307810 0.5431798 0.6405897 1.1805560 0.7109824
#> 16 1.6198356 2.0350829 2.3697261 2.6105717 2.7532895 1.1440421 1.6403368
#> 17 1.1191433 1.5308558 1.8637863 2.1035914 2.2454621 0.6465939 1.1313242
#> 18 0.6066502 0.9720382 1.2886473 1.5208227 1.6584865 0.3340139 0.5871279
#> 19 0.5355419 0.6647764 0.9043195 1.1066602 1.2308801 0.7094871 0.4686404
#> 20 0.7185440 0.6558247 0.7819527 0.9369336 1.0410094 1.0013910 0.6497947
#> 21 1.9514235 2.3666982 2.7013506 2.9422059 3.0849409 1.4757835 1.9721472
#> 22 1.5225570 1.9352799 2.2685419 2.5084839 2.6504232 1.0478102 1.5361383
#> 23 1.0014787 1.3872149 1.7081327 1.9415144 2.0795616 0.5897082 0.9934187
#> 24 0.7560457 1.0137385 1.2829400 1.4922112 1.6182093 0.6599905 0.7076087
#> 25 0.8311632 0.9267899 1.1158351 1.2876682 1.3960172 0.9357072 0.7647407
#>            8         9        10        11        12        13        14
#> 1  0.4685045 0.7755702 0.9500087 1.0855087 0.5580196 0.2481909 0.5429496
#> 2  0.1197270 0.3778043 0.5488347 1.5006379 0.9665308 0.4757039 0.3489695
#> 3  0.3163142 0.1551069 0.2600959 1.8352388 1.2988762 0.7810246 0.4830211
#> 4  0.5477014 0.2768914 0.1933986 2.0760548 1.5385417 1.0118296 0.6713901
#> 5  0.6871815 0.4013983 0.2689920 2.2187379 1.6803898 1.1494993 0.7942539
#> 6  0.9339559 1.2418792 1.4158657 0.6094186 0.1237082 0.5323894 0.9474151
#> 7  0.4365580 0.7446475 0.9188542 1.1055983 0.5664829 0.1818911 0.4859885
#> 8  0.0000000 0.3081702 0.4825549 1.5410538 0.9968816 0.4651961 0.2390345
#> 9  0.3081702 0.0000000 0.1745244 1.8481957 1.3022086 0.7562767 0.3949321
#> 10 0.4825549 0.1745244 0.0000000 2.0215172 1.4746130 0.9242080 0.5406423
#> 11 1.5410538 1.8481957 2.0215172 0.0000000 0.5503412 1.1136304 1.5320864
#> 12 0.9968816 1.3022086 1.4746130 0.5503412 0.0000000 0.5635108 0.9822899
#> 13 0.4651961 0.7562767 0.9242080 1.1136304 0.5635108 0.0000000 0.4189417
#> 14 0.2390345 0.3949321 0.5406423 1.5320864 0.9822899 0.4189417 0.0000000
#> 15 0.3552280 0.2879787 0.3718907 1.7646498 1.2152160 0.6521594 0.2334166
#> 16 2.0755685 2.3824402 2.5555103 0.5347813 1.0818502 1.6428769 2.0596980
#> 17 1.5615624 1.8658536 2.0374358 0.1328793 0.5651623 1.1178992 1.5321830
#> 18 0.9737492 1.2638216 1.4292019 0.6787723 0.2361702 0.5091695 0.9029725
#> 19 0.6051559 0.8311415 0.9752314 1.2111526 0.6922269 0.2873587 0.4368762
#> 20 0.5533010 0.6662851 0.7729465 1.5223266 0.9969319 0.5008301 0.3165364
#> 21 2.4073244 2.7141020 2.8870746 0.8665997 1.4130205 1.9729935 2.3889581
#> 22 1.9666906 2.2707653 2.4420942 0.4514726 0.9702594 1.5211287 1.9331664
#> 23 1.3941736 1.6852473 1.8502024 0.4351376 0.4704443 0.9303860 1.3207863
#> 24 0.9741682 1.2179360 1.3635852 0.9766058 0.5779279 0.5524468 0.8248469
#> 25 0.8471347 1.0135567 1.1289499 1.3461383 0.8873986 0.5832554 0.6340063
#>           15        16        17        18        19        20        21
#> 1  0.7630428 1.6198356 1.1191433 0.6066502 0.5355419 0.7185440 1.9514235
#> 2  0.4729018 2.0350829 1.5308558 0.9720382 0.6647764 0.6558247 2.3666982
#> 3  0.4307810 2.3697261 1.8637863 1.2886473 0.9043195 0.7819527 2.7013506
#> 4  0.5431798 2.6105717 2.1035914 1.5208227 1.1066602 0.9369336 2.9422059
#> 5  0.6405897 2.7532895 2.2454621 1.6584865 1.2308801 1.0410094 3.0849409
#> 6  1.1805560 1.1440421 0.6465939 0.3340139 0.7094871 1.0013910 1.4757835
#> 7  0.7109824 1.6403368 1.1313242 0.5871279 0.4686404 0.6497947 1.9721472
#> 8  0.3552280 2.0755685 1.5615624 0.9737492 0.6051559 0.5533010 2.4073244
#> 9  0.2879787 2.3824402 1.8658536 1.2638216 0.8311415 0.6662851 2.7141020
#> 10 0.3718907 2.5555103 2.0374358 1.4292019 0.9752314 0.7729465 2.8870746
#> 11 1.7646498 0.5347813 0.1328793 0.6787723 1.2111526 1.5223266 0.8665997
#> 12 1.2152160 1.0818502 0.5651623 0.2361702 0.6922269 0.9969319 1.4130205
#> 13 0.6521594 1.6428769 1.1178992 0.5091695 0.2873587 0.5008301 1.9729935
#> 14 0.2334166 2.0596980 1.5321830 0.9029725 0.4368762 0.3165364 2.3889581
#> 15 0.0000000 2.2910484 1.7624055 1.1276446 0.6299660 0.4015029 2.6196930
#> 16 2.2910484 0.0000000 0.5317366 1.1791411 1.7151687 2.0259667 0.3318251
#> 17 1.7624055 0.5317366 0.0000000 0.6475990 1.1838980 1.4949896 0.8575581
#> 18 1.1276446 1.1791411 0.6475990 0.0000000 0.5364375 0.8477736 1.5020818
#> 19 0.6299660 1.7151687 1.1838980 0.5364375 0.0000000 0.3115072 2.0363714
#> 20 0.4015029 2.0259667 1.4949896 0.8477736 0.3115072 0.0000000 2.3460059
#> 21 2.6196930 0.3318251 0.8575581 1.5020818 2.0363714 2.3460059 0.0000000
#> 22 2.1619408 0.1905093 0.4051727 1.0380735 1.5698392 1.8784945 0.4697213
#> 23 1.5415232 0.8222216 0.3333724 0.4215450 0.9281774 1.2316999 1.1258177
#> 24 1.0147385 1.4292064 0.9146274 0.3445777 0.3886223 0.6537837 1.7345857
#> 25 0.7584910 1.8093125 1.2916634 0.6796670 0.2968216 0.3576632 2.1146418
#>           22        23        24        25
#> 1  1.5225570 1.0014787 0.7560457 0.8311632
#> 2  1.9352799 1.3872149 1.0137385 0.9267899
#> 3  2.2685419 1.7081327 1.2829400 1.1158351
#> 4  2.5084839 1.9415144 1.4922112 1.2876682
#> 5  2.6504232 2.0795616 1.6182093 1.3960172
#> 6  1.0478102 0.5897082 0.6599905 0.9357072
#> 7  1.5361383 0.9934187 0.7076087 0.7647407
#> 8  1.9666906 1.3941736 0.9741682 0.8471347
#> 9  2.2707653 1.6852473 1.2179360 1.0135567
#> 10 2.4420942 1.8502024 1.3635852 1.1289499
#> 11 0.4514726 0.4351376 0.9766058 1.3461383
#> 12 0.9702594 0.4704443 0.5779279 0.8873986
#> 13 1.5211287 0.9303860 0.5524468 0.5832554
#> 14 1.9331664 1.3207863 0.8248469 0.6340063
#> 15 2.1619408 1.5415232 1.0147385 0.7584910
#> 16 0.1905093 0.8222216 1.4292064 1.8093125
#> 17 0.4051727 0.3333724 0.9146274 1.2916634
#> 18 1.0380735 0.4215450 0.3445777 0.6796670
#> 19 1.5698392 0.9281774 0.3886223 0.2968216
#> 20 1.8784945 1.2316999 0.6537837 0.3576632
#> 21 0.4697213 1.1258177 1.7345857 2.1146418
#> 22 0.0000000 0.6562125 1.2651242 1.6453118
#> 23 0.6562125 0.0000000 0.6089785 0.9892655
#> 24 1.2651242 0.6089785 0.0000000 0.3803503
#> 25 1.6453118 0.9892655 0.3803503 0.0000000