Graphical Web User Interface for SOMbrero

Start the SOMbrero GUI.

sombreroGUI()

Value

This function starts the graphical user interface with the default system browser. This interface is more lickely to work properly with Firefox https://www.mozilla.org/fr/firefox/new/. In case Firefox is not your default browser, copy/paste http://localhost:8100 into the URL bar.

References

Villa-Vialaneix N. (2017) Stochastic self-organizing map variants with the R package SOMbrero. In: J.C. Lamirel, M. Cottrell, M. Olteanu, 12th International Workshop on Self-Organizing Maps and Learning Vector Quantization, Clustering and Data Visualization (Proceedings of WSOM 2017), IEEE, Nancy, France.

RStudio and Inc. (2013). shiny: Web Application Framework for R. R package version 0.7.0. https://cran.r-project.org/package=shiny

Author

Élise Maigné <elise.maigne@inrae.fr>
Julien Boelaert julien.boelaert@gmail.com
Madalina Olteanu olteanu@ceremade.dauphine.fr
Nathalie Vialaneix nathalie.vialaneix@inrae.fr