2002 French presidential election data set

This data set provides the number of votes at the first round of the 2002 French presidential election for each of the 16 candidates for 106 administrative districts called "Départements".

Format

presidentielles2002 is a data frame of 106 rows (the French administrative districts called "Départements") and 16 columns (the candidates).

Source

The data are provided by the French ministry "Ministère de l'Intérieur". The original data can be downloaded at https://www.interieur.gouv.fr/Elections/Les-resultats/Presidentielles (2002 élections and "Résultats par départements").

References

The 2002 French presidential election consisted of two rounds. The second round attracted a greater than usual amount of international attention because of far-right candidate Le Pen's unexpected victory over Socialist candidate Lionel Jospin. The event is known because, on the one hand, the number of candidates was unusually high (16) and, on the other hand, because the polls had failed to predict that Jean-Marie Le Pen would be on the second round.

Further comments at https://en.wikipedia.org/wiki/2002_French_presidential_election.

Examples

data(presidentielles2002)
apply(presidentielles2002, 2, sum)
#>      MEGRET      LEPAGE  GLUCKSTEIN      BAYROU      CHIRAC      LE_PEN 
#>      667043      535875      132696     1949219     5666021     4804772 
#>     TAUBIRA SAINT_JOSSE      MAMERE      JOSPIN      BOUTIN         HUE 
#>      660515     1204801     1495774     4610267      339157      960548 
#> CHEVENEMENT     MADELIN   LAGUILLER  BESANCENOT 
#>     1518568     1113551     1630118     1210562